pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ayaka_who_is_suspect"
version = "0.0.5"
description = "谁是卧底"
authors = ["Su <wxlxy316@163.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bridgeL/nonebot-plugin-ayaka-who-is-suspect"
packages = [
    { include = "ayaka_who_is_suspect.py", from = "." },
]


[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b5"
nonebot-adapter-onebot = "^2.1.3"
nonebot-plugin-ayaka = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
